+++ metron-platform/metron-common/src/main/java/org/apache/metron/common/utils/ErrorUtils.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.metron.storm.common.utils;
+package org.apache.metron.common.utils;
 
 import static java.lang.String.format;
 
@@ -26,10 +26,6 @@ import java.lang.management.ThreadMXBean;
 import java.util.Optional;
 import java.util.function.Function;
 import org.apache.commons.lang3.tuple.Pair;
-import org.apache.metron.common.Constants;
-import org.apache.metron.common.error.MetronError;
-import org.apache.storm.task.OutputCollector;
-import org.apache.storm.tuple.Values;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -87,22 +83,6 @@ public class ErrorUtils {
         return format("%s", reason);
       }
     }
-  }
-
-  /**
-   * Handles a {@link MetronError} that occurs.
-   *
-   * @param collector The Storm output collector being reported to
-   * @param error The error that occurred
-   */
-  public static void handleError(OutputCollector collector, MetronError error)
-  {
-    collector.emit(Constants.ERROR_STREAM, new Values(error.getJSONObject()));
-    Optional<Throwable> throwable = error.getThrowable();
-    if (throwable.isPresent()) {
-      collector.reportError(throwable.get());
-    }
-
   }
 
   /**
