+++ metron-platform/metron-enrichment/metron-enrichment-storm/src/main/java/org/apache/metron/enrichment/bolt/UnifiedEnrichmentBolt.java
@@ -27,7 +27,6 @@ import org.apache.metron.common.error.MetronError;
 import org.apache.metron.storm.common.message.MessageGetStrategy;
 import org.apache.metron.storm.common.message.MessageGetters;
 import org.apache.metron.common.performance.PerformanceLogger;
-import org.apache.metron.storm.common.utils.ErrorUtils;
 import org.apache.metron.common.utils.MessageUtils;
 import org.apache.metron.enrichment.adapters.maxmind.asn.GeoLiteAsnDatabase;
 import org.apache.metron.enrichment.adapters.maxmind.geo.GeoLiteCityDatabase;
@@ -41,6 +40,7 @@ import org.apache.metron.enrichment.parallel.ConcurrencyContext;
 import org.apache.metron.enrichment.parallel.WorkerPoolStrategies;
 import org.apache.metron.stellar.dsl.Context;
 import org.apache.metron.stellar.dsl.StellarFunctions;
+import org.apache.metron.storm.common.utils.StormErrorUtils;
 import org.apache.storm.task.OutputCollector;
 import org.apache.storm.task.TopologyContext;
 import org.apache.storm.topology.OutputFieldsDeclarer;
@@ -285,7 +285,7 @@ public class UnifiedEnrichmentBolt extends ConfiguredEnrichmentBolt {
                 .withMessage(t.getValue().getMessage())
                 .withThrowable(t.getValue())
                 .addRawMessage(t.getKey());
-        ErrorUtils.handleError(collector, error);
+        StormErrorUtils.handleError(collector, error);
       }
     } catch (Exception e) {
       //If something terrible and unexpected happens then we want to send an error along, but this
@@ -296,7 +296,7 @@ public class UnifiedEnrichmentBolt extends ConfiguredEnrichmentBolt {
               .withMessage(e.getMessage())
               .withThrowable(e)
               .addRawMessage(message);
-      ErrorUtils.handleError(collector, error);
+      StormErrorUtils.handleError(collector, error);
     }
     finally {
       collector.ack(input);
