+++ metron-platform/metron-writer/src/test/java/org/apache/metron/writer/BulkWriterComponentTest.java
@@ -33,7 +33,7 @@ import java.util.Collections;
 import java.util.List;
 
 import org.apache.metron.common.configuration.writer.WriterConfiguration;
-import org.apache.metron.storm.common.utils.ErrorUtils;
+import org.apache.metron.common.utils.ErrorUtils;
 import org.apache.metron.common.writer.BulkMessageWriter;
 import org.apache.metron.common.writer.BulkMessage;
 import org.apache.metron.common.writer.BulkWriterResponse;
